+++ src/neo-express/Commands/ContractCommand.cs
@@ -1,4 +1,5 @@
-﻿using McMaster.Extensions.CommandLineUtils;
+﻿using System;
+using McMaster.Extensions.CommandLineUtils;
 
 namespace NeoExpress.Commands
 {
@@ -11,6 +12,22 @@ namespace NeoExpress.Commands
         typeof(Storage))]
     internal partial class ContractCommand
     {
+
+        static string GetScriptHash(string contract)
+        {
+            if (Neo.UInt160.TryParse(contract, out var _))
+            {
+                return contract;
+            }
+
+            if (BlockchainOperations.TryLoadContract(contract, out var _contract, out var errorMessage))
+            {
+                return _contract.Hash;
+            }
+
+            throw new Exception(errorMessage);
+        }
+
         private int OnExecute(CommandLineApplication app, IConsole console)
         {
             console.WriteError("You must specify at a subcommand.");