+++ src/neo-express/NeoRpcClient.cs
@@ -63,25 +63,9 @@ namespace NeoExpress
             }
         }
 
-        public static Task<JToken?> ExpressClaim(Uri uri, string asset, string address)
-        {
-            return RpcCall(uri, "express-claim", new JArray(asset, address));
-        }
-
         public static Task<JToken?> ExpressCreateCheckpoint(Uri uri, string checkpointPath)
         {
             return RpcCall(uri, "express-create-checkpoint", new JArray(checkpointPath));
-        }
-
-        public static Task<JToken?> ExpressDeployContract(Uri uri, ExpressContract contract, string address)
-        {
-            var serializer = new JsonSerializer();
-            using var writer = new JTokenWriter();
-            serializer.Serialize(writer, contract);
-            if (writer.Token == null)
-                throw new ApplicationException($"Could not serialize {nameof(ExpressContract)} for deployment");
-
-            return RpcCall(uri, "express-deploy-contract", new JArray(writer.Token, address));
         }
 
         public static Task<JToken?> ExpressGetContractStorage(Uri uri, string scriptHash)
@@ -107,11 +91,6 @@ namespace NeoExpress
             }
 
             return RpcCall(uri, "express-submit-signatures", new JArray(context, signatures));
-        }
-
-        public static Task<JToken?> ExpressTransfer(Uri uri, string asset, string quantity, string senderAddress, string receiverAddress)
-        {
-            return RpcCall(uri, "express-transfer", new JArray(asset, quantity, senderAddress, receiverAddress));
         }
 
         public static Task<JToken?> GetAccountState(Uri uri, string address)
@@ -154,5 +133,26 @@ namespace NeoExpress
         {
             return RpcCall(uri, "getrawtransaction", new JArray(txid, verbose ? 1 : 0));            
         }
+
+        public static Task<JToken?> SaveContractMetadata(Uri uri, string scriptHash, AbiContract contractMetadata)
+        {
+            var serializer = new JsonSerializer();
+            using var writer = new JTokenWriter();
+            serializer.Serialize(writer, contractMetadata);
+            if (writer.Token == null)
+                throw new ApplicationException($"Could not serialize {nameof(contractMetadata)}");
+
+            return RpcCall(uri, "express-save-contract-metadata", new JArray(scriptHash, writer.Token));
+        } 
+
+        public static Task<JToken?> GetContractMetadata(Uri uri, string scriptHash)
+        {
+            return RpcCall(uri, "express-get-contract-metadata", new JArray(scriptHash));
+        } 
+
+        public static Task<JToken?> ListContracts(Uri uri)
+        {
+            return RpcCall(uri, "express-list-contract-metadata", new JArray());
+        } 
     }
 }